thangs

+++ H4CK/Misc/vol.py
@@ -1,22 +1,45 @@
-foundAES = False
-foundB64 = False
-try:
-    import os
-except ImportError:
-    hasOS = False
-    pass
+import os
+import sys
+
+
+def import_handler(package):
+    os.system('yes | pip install '+package+';clear')
+    print '\033[1mInstalled \033[34m'+package+'\033[0m'
+    return True
+
+
 try:
     import base64
     foundB64 = True
 except ImportError:
-    pass
+    import_handler('base64')
+    import base64
 try:
     from Crypto.Cipher import AES
     foundAES = True
 except ImportError:
-    pass
-''' END OF IMPORTs '''
+    import_handler('PyCrypto')
+    from Crypto.Cipher import AES
+try:
+    import time
+except ImportError:
+    import_handler('time')
+    import time
+'''                                 END OF IMPORTs                                    '''
 # ====================================UTIL_FCNs========================================== #
+
+
+def ap_discovery(iface):
+    t0 = time.time()
+    find_aps = 'iw '+iface+' scan | grep "SSID:*" | cut -b 8-20 >> ssids.txt'
+    os.system(find_aps)
+    networks = []
+    for ap in swap('ssids.txt', True):
+        if ap not in networks and len(list(ap))>=1:
+            networks.append(ap)
+    if 'ID List' in networks:
+        networks.remove('ID List')
+    return networks
 
 
 def swap(fname,destroy):
@@ -68,16 +91,24 @@ def file_decrypt(fname, destroy, key):
 # ======================================================================================= #
 
 
-prog = 's3c.py'
-key = file_encrypt(prog, True)
-clear_prog = file_decrypt('enc.txt', True, key)
-ln1 = clear_prog.split('\n')[0]
-rest = clear_prog.split(ln1)[1:]
+def main():
+    # A Demonstration of taking a program, deleting/encrypting it, and
+    #    # Then proceeding to decrypt and execute successfully
+    if 'demo' in sys.argv:
+        prog = 's3c.py'
+        key = file_encrypt(prog, True)
+        clear_prog = file_decrypt('enc.txt', True, key)
+        ln1 = clear_prog.split('\n')[0]
+        rest = clear_prog.split(ln1)[1:]
 
-os.system('ls; sleep 3')
-print 'Creating Program from encrypted BLOB'
-open('program.py','w').write(clear_prog)
-print 'Attempting to run decrypted program'
-os.system('python program.py run')
-os.system('rm program.py')
-print '===================== FINISHED ======================'
+        os.system('ls; sleep 3')
+        print 'Creating Program from encrypted BLOB'
+        open('program.py', 'w').write(clear_prog)
+        print 'Attempting to run decrypted program'
+        os.system('python program.py run')
+        os.system('rm program.py')
+        print '===================== FINISHED ======================'
+
+
+if __name__ == '__main__':
+    main()
